refactor: modularize database migration code
